requirements.txt
django
djangorestframework
django-location-field
Pillow
django-rest-auth
django-allauth
django-geoposition

